Generated by Spark: Hosted by ELENA 9.8 Host Review Score
ELENA
Located in a private and wooded domain, a magnificent villa recently renovated with infinity pool, sea and Monaco view, 15 minutes of driving from Monte Carlo, Nice and Italy. Ideal for holidays with family, friends or for business stay. 1 master suite with bathtub and private terrace upstairs, 2 bedrooms with independent italian showers and 1 TV room with shower convertible into a bedroom on request, all sea and gardern view. Spacious 60m2 living and dining area with all-equipped modern kitchen and fireplace. Large swimming pool with 3 depths for all ages, a shady area for a moment of relaxation, gas BBQ, summer lounge, lush garden. Private outddor parking for 2 cars. An adorable guest house with small kitchen for 4 people can be booked additionally with the main property.
I speak french, english, italiand and russian and it's my pleasure to offer you an unforgettable stay in my beautiful country. There are so many things to do and to see here! I'll try to be as helpful as possible during your stay. Looking forward to host you at any moment of the year!
La Turbie is a small town situated above the Monaco principality on the French Riviera and the spectacular Grand Corniche road that runs from Nice to Monaco - one of the great scenic highlights of this region. The medieval village in La Turbie is small but nicely preserved and it is the great Roman monument in La Turbie that attracts visitors - it dominates the village from its high position. As for an everyday life, you will find all you can need in the village: restaurants, pharmacy, bakery, cheese, meat and grocery stores, hairdresser, tourism office... There are too many highlights in the region to list here, from magnificent scenery and traditional Provencal hilltop villages such as Eze to the glitzy resorts of the French riviera like Roquebrune Cap Martin, Beaulieu sur Mer, Saint Jean Cap Ferrat and the wealthy principality of Monaco. Nice airport and Italy are also in a quarter of hour of drive from la Turbie. There is something close to La Turbie for everyone!
Languages spoken:
English,French,Italian,Russian

and use elena.jpg add hosted by Elena for Villa Azure